High or Low game
This Python code is a simple game that allows users to compare the number of followers or subscribers of different celebrities. The game randomly selects two celebrities from a predefined dictionary called "celebs," which contains the names of various celebrities as keys and their respective follower counts as values. The user is then prompted to choose which of the two celebrities has more followers by entering 'a' or 'b'.